카테고리 사이드바 수정
