+++ frontend/src/app/page.tsx
@@ -1,7 +1,8 @@
-import { getProductRecentList } from "@/app/API/NonUserAPI";
+import { getCategories, getProductRecentList } from "@/app/API/NonUserAPI";
 import Page from "./page_csr";
 
 export default async function Home() {
   const productList = await getProductRecentList(0);
-  return <Page productList={productList} />;
+  const categories = await getCategories();
+  return <Page productList={productList} categories={categories} />;
 }
