+++ frontend/src/app/page_csr.tsx
@@ -6,6 +6,7 @@ import { MonthDate } from './Global/Method';
 
 interface pageProps {
     productList: any;
+    categories: any[];
 }
 export default function Page(props: pageProps) {
     const [user, setUser] = useState(null as any);
@@ -24,7 +25,7 @@ export default function Page(props: pageProps) {
                 .catch(e => console.log(e));
     }, [ACCESS_TOKEN]);
 
-    return <Main user={user} recentList={recentList} setRecentList={setRecentList}>
+    return <Main user={user} recentList={recentList} setRecentList={setRecentList} categories={props.categories}>
         <div className='w-full h-full flex justify-center'>
             <div className='flex flex-wrap w-[1240px]'>
                 {(productList.content as any[]).map((product, index) =>
