+++ frontend/src/app/product/best/page.tsx
@@ -1,9 +1,10 @@
-import { getProductBest } from "@/app/API/NonUserAPI";
+import { getCategories, getProductBest } from "@/app/API/NonUserAPI";
 import Page from "./page_csr";
+import { getCartList } from "@/app/API/UserAPI";
 
 
 export default async function Home({ params }: { params: any }) {
     const bestList = await getProductBest();
-    
-    return <Page bestList={bestList}></Page>;
+    const categories = await getCategories();
+    return <Page bestList={bestList} categories={categories} />;
 }
