+++ frontend/src/app/product/best/page_csr.tsx
@@ -5,6 +5,7 @@ import { getRecent, getUser } from '@/app/API/UserAPI';
 
 interface pageProps {
     bestList: any[];
+    categories: any[];
 }
 
 export default function Page(props: pageProps) {
@@ -24,7 +25,7 @@ export default function Page(props: pageProps) {
                 .catch(e => console.log(e));
     }, [ACCESS_TOKEN]);
 
-    return <Main user={user} recentList={recentList} setRecentList={setRecentList}>
+    return <Main user={user} recentList={recentList} setRecentList={setRecentList} categories={props.categories}>
         <div className='w-full h-full flex justify-center'>
             <div className='flex flex-wrap w-[1240px]'>
                 {bestList.map((product, index) =>
